Создание модульности в js
