+++ js/photo-data-const.js
@@ -1,0 +1,42 @@
+const NAMES = [
+  'Ольга',
+  'Тимур',
+  'Владислав',
+  'Сергей',
+  'Юлия',
+  'Наталья',
+  'Игорь',
+  'Татьяна',
+  'Александр',
+  'Евгения',
+  'Дарья'
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTION = [
+  'Мужчина играет с собакой.',
+  'Цветущее поле в солнечный день.',
+  'Проливной дождь в городе.',
+  'Девушка читает книгу.',
+  'Девушка собирает цветы в поле.',
+  'Дети лепят снеговика.',
+  'Тигрица вывела тигрят погулять.',
+  'Красиво падает снег в ночном городе.',
+  'В бассейне купаются дети.',
+  'Кот сидит на подоконнике и смотрит в окно.'
+];
+
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MIN_AVATAR = 1;
+const MAX_AVATAR = 6;
+
+export { NAMES, MESSAGE, DESCRIPTION, MIN_LIKES, MAX_LIKES, MIN_AVATAR, MAX_AVATAR };
